Crate demo_vk

Modules

app

The Winit ApplicationHandler implementation.

graphics

Macros

here

Creates a “Location” type that displays the current location in code when printed.

unwrap_here

Tries to resolve a result into a successful output (much like the ? operator). Unlike the ? operator, this macro adds context about the current line to the error message if something goes wrongw.

Functions

map