Module demo_vk::graphics

Modules

vulkan

This module defines traits, structs, and functions to interact with a Vulkan device.

Structs

BindlessTextureAtlas

Maintains descriptors for a list of texture images. Textures can be accessed by their index in shaders when the descriptor is bound. e.g. “bindless” textures.

Recompiler

Automatically recompiles a slang source file any time it changes.

Sprite

A sprite is a textured quad that can be rendered with a SpriteLayer.

SpriteLayer

A sprite layer contains all of the resources to render batches of Sprites.

StreamingSprites

A sprite-batch implementation that supports streaming new sprites every frame.

SwapchainColorPass

A utility for managing a Renderpass and Framebuffers for presenting to swapchain images.

Texture

A Texture is a Vulkan Image paired with backing GPU memory and metadata.

TextureLoader

Functions

ortho_projection