demo_vk::graphics

Module streaming_renderer

Mesh, Material, Texture, and pipeline resources for rendering geometry that’s expected to change every frame.

Structs

StreamingRenderer

A renderer optimized for streaming new vertex data to the GPU every frame.

Texture

A 2D image for use when rendering.

TextureAtlas

The TextureAtlas maintains the per-frame descriptor sets used for binding textures.

TextureLoader

A utility for loading textures from image files.

TrianglesMesh

This mesh supports drawing arbitrary triangles and quads in three dimensions.

Vertex